// src/Models/contact.ts
import { CdkStepperModule } from '@angular/cdk/stepper';
export interface Contact {
    id?: string;
    identification: string;
    name: string;
    firstName: string;
    lastName: string;
    password: string;
    address: string;
    city: string;
    state: string;
    postalCode: string;
    email:string;
    phoneNumber: string;
    status: string;
    isUser: string;
    userId: string;
    companyid: string;

  }
  